eth2util: refactor project structure
