+++ dkg/disk.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/obolnetwork/charon/app/errors"
 	"github.com/obolnetwork/charon/cluster"
+	"github.com/obolnetwork/charon/eth2util/keystore"
 	"github.com/obolnetwork/charon/tbls/tblsconv"
-	"github.com/obolnetwork/charon/testutil/keystore"
 )
 
 // loadDefinition returns the cluster definition from disk (or the test definition if configured).
