+++ eth2util/keystore/testdata/keystore-scrypt.json
